merging impulse branch to collision branch

+++ the_engine/the_engine/src/shapes/triangle.cpp
@@ -6,7 +6,7 @@ Triangle::Triangle(float width, float height) {
 	t_height = height;
 };
 
-void Triangle::create_triangle(double deltat) {
+void Triangle::create_triangle(float deltat) {
 	//aspect ratio compensation
 	float new_height = (t_height / 9) / 2;
 	float new_width = (t_width / 16) / 2;
